Enable multiple SyncedFolder resources in programs
Fixes a bug in registering child resources of the SyncedFolder component which
manifests to users as an error:

Handling "Duplicate resource URN" and "Duplicate resource alias"
=============

If you see thes errors like "Duplicate resource alias ...", check if you are
creating more than one folder object of S3BucketFolder, AzureBlobFolder, or
GoogleCloudFolder.

On every resource except the first, you should add an input to the resource
`disableManagedObjectAliases`. That may look something like this, using
TypeScript as an example:

```ts
const folder = new synced.S3BucketFolder("synced-folder", {
    path: "./my-folder",
    bucketName: bucket.bucket,
    acl: aws.s3.PublicReadAcl,
    disableManagedObjectAliases: true,
});
```

This will allow your program to complete, and you should see output like the
following in a `pulumi preview`, with no deletes or replaces:

```
     Type                                   Name                        Plan
     pulumi:pulumi:Stack                    synced-folder-aws-yaml-dev
 +   ├─ aws:s3:Bucket                       my-bucket-2                 create
 +   └─ synced-folder:index:S3BucketFolder  synced-folder-2             create
 +      └─ aws:s3:BucketObject              synced-folder-2-index.html  create
```

If you see folders being deleted and aren't sure, reach out in an issue or on
our [Community Slack](https://slack.pulumi.com/) for support.

Future releases
=============

This field will be removed in v0.2.0 aafter a period of time to allow users to
migrate. Users should run a deployment using v0.1.0 **before** updating to
v0.2.0.

With the field `aliases` removed, users upgrading directly to v0.2.0 directly
will see Pulumi plan a "create" and "delete". Deploying on v0.1.0 will rename
the resources to prevent this.

Implementation notes
=============

This section is for pull request reviewers.

> error: Duplicate resource URN
> 'urn:pulumi:dev::synced-folder-aws-yaml::synced-folder:index:S3BucketFolder$aws:s3/bucketObject:BucketObject::index.html';
> try giving it a unique name

This error is caused by registering multiple child resources under the same
name, in this case `index.html`. Fixing this by only changing the resource name
however introduces a regression, users will would resource deletes and creates
like so:

```
     pulumi:pulumi:Stack                    synced-folder-aws-yaml-dev
     └─ synced-folder:index:S3BucketFolder  synced-folder
 +      ├─ aws:s3:BucketObject              synced-folder-index.html    create
 -      └─ aws:s3:BucketObject              index.html                  delete
```

To fix this, we use the [aliases resource
option](https://www.pulumi.com/docs/intro/concepts/resources/options/aliases/).
This ensuress the resource `synced-folder-index.html` is registered with an
alias for the old name `index.html`, which allows Pulumi to rename the URN
instead of planning a create and delete.

Users will encounter an error "Duplicate resource alias" if more than one
resource is registered, so an input `disableManagedObjectAliases` is added to
disabling the alias to the legacy name.